chore: move existing decorator and add a measure time decorator
